Add [[nodiscard]] attribute and update switch case type in matrix
Added [[nodiscard]] attribute to getter methods in matrix.hpp to suggest compiler that a method return value should not be ignored. The type of the switch case in the convertTileToArray method in matrix.cpp was updated from an unsigned short to std's underlying_type to better handle the enum class.