Improve CMake configuration and refactor matrix code
The CMake configuration was updated to include warnings and extra options, and was set to prevent in-source builds. Directory installation statements were added to CMakeLists.txt files in the top-level and src directories. For better maintenance, the functions 'mb' and 'nb' in the 'matrix' class have been moved from the header file to the implementation file.